feat(core)!: finalize the Core public API surface before the #6 freeze (#95)
Owner decisions from #63, applied as one pass so the PublicAPI baseline is
re-captured once.

- Remove unreachable public types: ChunkerId, ChunkerIds, ContentId,
  ManifestSchemaVersion, ManifestSchemaVersions. The RFC concepts stay.
- ChunkingProfileId and HashSuiteId become sealed, immutable reference
  value objects: Value is never null, IsDefault is gone, equality stays
  ordinal, and "no explicit selection" is a null option. Hash-suite
  resolution returns the canonical instance; resolution stays
  allocation-free (tested).
- Remove ChunkEntry. ManifestReader.ReadAsync fills Memory<ChunkInfo>, the
  shared logical chunk position model.
- Public integers are Int64 (Index, Offset, ChunkCount, ContentLength,
  PhysicalLength, ChunkBlockCount) and Int32 (Length). The wire format
  stays UInt64/UInt32. A well-formed manifest outside that range is
  NotSupportedException. The check sits in the shared CSM parser, so
  VerifyManifestAsync, VerifyAsync and ManifestReader reach the same
  verdict; the independent decoder models the same rule, and a new
  conformance vector covers a 2^31-byte chunk. No persisted byte,
  ManifestId or FileDigest changes.
- ChunkManifest.CreateAsync/VerifyAsync/VerifyManifestAsync and
  ManifestReader.OpenAsync/ReadAsync validate arguments synchronously
  with public parameter names, then run a private async core.
- Document ArgumentNull/Argument/InvalidData/NotSupported/
  OperationCanceled/InvalidOperation on every public operation.
- Re-capture PublicAPI.Unshipped.txt (171 -> 123 entries); Shipped stays
  empty.

BREAKING CHANGE: pre-release public API changes listed above.

Refs #63, #6

+++ tests/ChunkShift.Tests/Manifest/CsmIndependentVectorTests.cs
@@ -48,8 +48,8 @@ public sealed class CsmIndependentVectorTests
         byte[] bytes = File.ReadAllBytes(FixturePath(name));
 
         Assert.Equal(
-            vector.GetProperty("physicalLength").GetUInt64(),
-            (ulong)bytes.LongLength);
+            vector.GetProperty("physicalLength").GetInt64(),
+            bytes.LongLength);
 
         Task<ManifestVerificationResult> verification =
             ChunkManifest.VerifyManifestAsync(
@@ -83,19 +83,19 @@ public sealed class CsmIndependentVectorTests
 
         ManifestInfo manifest = result.Manifest;
         Assert.Equal(
-            vector.GetProperty("chunkCount").GetUInt64(),
+            vector.GetProperty("chunkCount").GetInt64(),
             manifest.ChunkCount);
         Assert.Equal(
-            vector.GetProperty("contentLength").GetUInt64(),
+            vector.GetProperty("contentLength").GetInt64(),
             manifest.ContentLength);
         Assert.Equal(
-            vector.GetProperty("cblkCount").GetUInt64(),
+            vector.GetProperty("cblkCount").GetInt64(),
             manifest.ChunkBlockCount);
         Assert.Equal(
             vector.GetProperty("hasBidx").GetBoolean(),
             manifest.HasBlockIndex);
         Assert.Equal(
-            vector.GetProperty("physicalLength").GetUInt64(),
+            vector.GetProperty("physicalLength").GetInt64(),
             manifest.PhysicalLength);
 
         if (outcome == "valid")
